Moved CLD to View package #2
